requirements.txt
# Web Framework
Flask
Flask-SQLAlchemy
Flask-Cors
Werkzeug

# Data Processing
pandas
numpy

# Machine Learning & NLP
torch
transformers
sentence-transformers
bertopic
spacy
nltk

# Additional ML libraries
scikit-learn
huggingface-hub

# API Integration
langchain_groq
groq
langchain

# Utilities
python-dotenv

# Visualization (if needed)
matplotlib
seaborn
wordcloud
umap-learn